Move VaccineCardComponent to components/public/vaccine-card/
