+++ Apps/WebClient/src/NewClientApp/src/components/public/vaccine-card/VaccineCardComponent.vue
