+++ Apps/WebClient/src/NewClientApp/src/components/public/vaccine-card/VaccineCardView.vue
@@ -11,8 +11,8 @@ import HgButtonComponent from "@/components/common/HgButtonComponent.vue";
 import HgDatePickerComponent from "@/components/common/HgDatePickerComponent.vue";
 import LoadingComponent from "@/components/common/LoadingComponent.vue";
 import MessageModalComponent from "@/components/common/MessageModalComponent.vue";
-import VaccineCardComponent from "@/components/common/VaccineCardComponent.vue";
 import TooManyRequestsComponent from "@/components/error/TooManyRequestsComponent.vue";
+import VaccineCardComponent from "@/components/public/vaccine-card/VaccineCardComponent.vue";
 import { container } from "@/ioc/container";
 import { SERVICE_IDENTIFIER } from "@/ioc/identifier";
 import CovidVaccineRecord from "@/models/covidVaccineRecord";
